Refactored code structure
