--- Components/Elements/Interface/Forms/index.tsx
@@ -1,1 +1,0 @@
-export * from "./StyledTextInput"
