--- Components/Elements/Interface/index.tsx
@@ -1,4 +1,0 @@
-export * from "./BigButton"
-export * from "./Card"
-export * from "./SearchBox"
-export * from "./CardsScroll"
