+++ ui/text/Price.tsx
@@ -1,13 +1,13 @@
 import { StyleSheet, Text, View, StyleProp, TextStyle } from "react-native";
-import React, { ReactNode } from "react";
-import { ColorPalette } from "../../../assets/Globals";
+import { ReactNode } from "react";
+import { ColorPalette } from "../../constants";
 
 interface PriceProps {
   style?: StyleProp<TextStyle>;
   children?: ReactNode;
 }
 
-export function Price({style, children}: PriceProps) {
+export function Price({ style, children }: PriceProps) {
   return <Text style={[styles.header, style]}>{children}</Text>;
 }
